+++ CHANGES.rst
@@ -1,11 +1,30 @@
 Changelog
 =========
 
-5.4.9 (unreleased)
+5.4.10 (unreleased)
+-------------------
+
+- Fixed limit event occurrences to 100.
+  [eikichi18]
+
+
+5.4.9 (2024-04-22)
 ------------------
 
-- Nothing changed yet.
-
+- Limit event occurrences to 100.
+  [mamico]
+- Customize INameChooser adapter to check also alias ids and disallow to create contents that could override aliases.
+  [cekk]
+- Customize also `copy` and `move` endpoints to raise BadRequest if that action will override some aliases.
+  [cekk]
+- Add flag in controlpanel to enable/disable INameChooser customization.
+  [cekk]
+- Exclude bg_color from transformed fields in deserializer.
+  [cekk]
+- Uninstall collective.volto.cookieconsent (deprecated). Will be removed from dependencies in next releases.
+  [cekk]
+- Add dependency to collective.volto.gdprcookie and install it by default.
+  [cekk]
 
 5.4.8 (2024-03-19)
 ------------------